# 04_API/openapi.yaml
openapi: 3.0.3
info:
  title: AIM Tank Integrity API
  version: 0.10.0
  description: Current implemented AIM+n8n Tank Integrity endpoints through Sprint 10 report generation. Deterministic universal calculation, FFS/RBI trigger workflow, engineering review/approval, and draft DOCX/PDF report generation are implemented; no API/API-ASME formula expressions are hard-coded or proprietary quantitative API RP 581 rules embedded.
servers:
  - url: http://localhost:4000/api/v1
security:
  - DemoRoleHeaders: []
...
paths:
...
  /engineering/calculate:
    post:
      summary: Run deterministic engineering calculation
      description: Runs universal deterministic calculations only. API/API-ASME-dependent calculations must use approved Formula Registry metadata and are not hard-coded. Calculation blocks when validation severity is blocking.
      x-permission-required: calculation.run
      x-audit-event-generated: CALCULATION_RUN_CREATED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineeringCalculationRequest'
      responses:
        '201':
          description: Calculation completed and stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineeringCalculationResponse'
        '422':
          description: Calculation blocked by validation and run snapshot stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineeringCalculationResponse'
        '400':
          description: Invalid request, approved Formula Registry record missing, or non-universal deterministic formula blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /engineering/calculations:
    get:
      summary: List calculation runs
      x-permission-required: calculation.read
      parameters:
        - in: query
          name: asset_id
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Calculation run list }
...
  /engineering/calculations/{runId}:
    get:
      summary: Get calculation detail with inputs, outputs, engineering reviews, approvals, and audit trail
      x-permission-required: calculation.read
      parameters:
        - in: path
          name: runId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Calculation detail and audit trail }
        '404': { description: Calculation run not found }
...
components:
  securitySchemes:
    DemoRoleHeaders:
      type: apiKey
      in: header
      name: x-aim-demo-roles
...
  schemas:
...
    EngineeringCalculationRequest:
      type: object
      required: [asset_id]
      additionalProperties: true
      properties:
        asset_id:
          type: string
          format: uuid
        inspection_event_id:
          type: string
          format: uuid
        formula_id:
          type: string
          default: AIM-UNIVERSAL-THICKNESS-CORROSION-ENGINE
        formula_version:
          type: string
        calculation_scope:
          type: string
          enum: [thickness_screening, corrosion_rate, remaining_life, inspection_interval_placeholder]
        calculation_request:
          type: object
          properties:
            thickness_check_requested: { type: boolean }
            retirement_thickness_mm: { type: number }
            controlled_formula_output: { type: object }
        thresholds:
          type: object
          properties:
            high_corrosion_rate_mm_per_year: { type: number }
            low_remaining_life_years: { type: number }
        context:
          type: object
          description: Optional validation/calculation context override. Stored data is loaded by asset_id when omitted.
    EngineeringCalculationResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            run_id: { type: string }
            run_status: { type: string }
            validation_status: { type: string, enum: [passed, blocked] }
            formula_set_version: { type: string }
            input_snapshot_hash: { type: string }
            output_summary: { type: object }
            calculation:
              type: object
              properties:
                deterministic_engine_version: { type: string }
                corrosion_rates:
                  type: array
                  items: { type: object }
                remaining_life:
                  type: array
                  items: { type: object }
                warnings:
                  type: array
                  items: { type: object }
        auditLogId: { type: string }
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items: { type: object }


# Sprint 7 governance hardening notes
# Sprint 8.5 evidence linkage hardening: generic evidence_links rejects cross-asset links with CROSS_ASSET_EVIDENCE_LINK_BLOCKED.
# NDT approval cannot use linked evidence from another asset.
# FFS case creation rejects cross-asset evidence links with error code CROSS_ASSET_EVIDENCE_LINK_BLOCKED.
# Demo header authentication is local-development only; production deployments require verified JWT/session identity and DB-backed RBAC.

# Sprint 8 RBI interface governance notes
# RBI cases are qualitative/semi-quantitative placeholders unless approved quantitative Formula Registry rules are provided.
# No proprietary API RP 581 quantitative rules are embedded in this implementation.

# Sprint 9 engineering review and approval workflow governance notes
# Engineers may review calculation/data records; senior_engineer/admin approval is required for final approval, overrides, rejection, and locking.
# Override approval requires affected field/result, original value, override value, reason, evidence, approver, and audit timestamp. AI agents cannot approve, reject, override, or lock engineering decisions.


# Sprint 10 report generation governance notes
# /reports/generate creates DOCX/PDF draft outputs from locked or review-ready calculation runs.
# Reports cite formula_id, formula_version, code_basis, code_edition, calculation_run_id, and input snapshot hash.
# Draft reports are explicitly marked DRAFT until approved; issued reports are locked and require a new version for changes.
# No API/API-ASME formula expression is embedded or invented in generated report content.